requirements.txt
ollama
discord.py
dotenv